Refs #50. Move LoadingExtension to separate class. Don't detour ImmaterialResourceManager if 81 Tiles mod is active.
