+++ BuildingThemes/BuildingThemes.csproj
@@ -75,6 +75,7 @@
     <Compile Include="GUI\ThemeManager\UIThemeManager.cs" />
     <Compile Include="GUI\UITitleBar.cs" />
     <Compile Include="GUI\UIUtils.cs" />
+    <Compile Include="LoadingExtension.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Detour\RedirectionHelper.cs" />
     <Compile Include="SerializableDataExtension.cs" />
